Sub: tweak Yaw behavior for workign at both high and low gains in an useable way

+++ ArduSub/Sub.h
@@ -411,7 +411,7 @@ private:
     void update_altitude();
     float get_smoothing_gain();
     void get_pilot_desired_lean_angles(float roll_in, float pitch_in, float &roll_out, float &pitch_out, float angle_max);
-    float get_pilot_desired_yaw_rate(int16_t stick_angle) const;
+    float get_pilot_desired_yaw_rate(float stick_angle) const;
     void check_ekf_yaw_reset();
     float get_roi_yaw();
     float get_look_ahead_yaw();
